requirements.txt
absl-py==0.13.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
docopt==0.6.2
gast==0.5.2
gitdb==4.0.7
GitPython==3.1.18
google-pasta==0.2.0
grpcio==1.39.0
h5py==2.10.0
idna==3.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.1
joblib==1.0.1
jsonpickle==1.5.2
Keras==2.1.5
Keras-Applications==1.0.8
keras-contrib @ git+https://www.github.com/farizrahman4u/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
munch==2.5.0
networkx==2.5.1
nibabel==3.2.1
nilearn==0.8.1
numexpr==2.7.3
numpy==1.19.5
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
Pillow==8.3.2
protobuf==3.17.3
py-cpuinfo==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
sacred==0.8.2
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.5.4
Shapely==1.7.1
SimpleITK==2.1.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
tables==3.6.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2020.9.3
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.5.0
